Another one
git-svn-id: file:///home/toad/git-migration/temprepository/branches/db4o/freenet@25387 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/support/io/SegmentedChainBucketSegment.java
@@ -81,6 +81,10 @@ public class SegmentedChainBucketSegment {
 	
 	public void removeFrom(ObjectContainer container) {
 		for(Bucket bucket : buckets) {
+			if(bucket == null) {
+				Logger.error(this, "Bucket is null in removeFrom() on "+this);
+				continue;
+			}
 			container.activate(bucket, 1);
 			bucket.removeFrom(container);
 		}
